pyproject.toml
[project]
name = "release-notes-generator"
version = "1.0.0"
description = "Keboola Custom Python Component for automatic generation of release notes"
authors = [
    {name = "Keboola", email = "support@keboola.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.27.0",
    "google-generativeai>=0.3.0",
    "keboola.component>=0.1.0",
    "pydantic>=2.0.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
] 
